bmsapi/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
firebase-admin==6.4.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0 